refact FilterScreen logic
