+++ lib/domain/FilterItem.dart
@@ -1,0 +1,7 @@
+class FilterItem {
+  String name = "";
+  String assetPath = "";
+  bool isSelected = true;
+
+  FilterItem(this.name, this.assetPath);
+}
